Backend/requirements.txt
redis==5.0.3
python-dotenv==1.0.1
flask
flask-cors
